refactor: split app.js into per-tab ES modules
Breaks the 2,800-line monolithic app.js into seven focused ES modules
under static/js/ with a thin entry-point app.js (~50 LOC). No behaviour
changes — pure structural refactor.

New layout:
  static/app.js           entry point (imports, subscriptions, boot)
  static/js/store.js      state object + reactive store
  static/js/utils.js      api, esc, toast, fmt*, oplog helpers
  static/js/loader.js     loadAll, SSE client, polling fallback
  static/js/pools.js      Sysinfo tab (pools, SMART, scrub, auto-snap)
  static/js/datasets.js   Datasets tab (CRUD, ACL, NFS, SMB, iSCSI)
  static/js/snapshots.js  Snapshots tab
  static/js/users.js      Users & Groups tab (users, groups, Samba, TM)

Uses native browser ES modules (type="module") — no build step.
