+++ static/index.html
@@ -891,6 +891,6 @@
   <!-- TOAST -->
   <div id="toast" class="toast" aria-live="polite"></div>
 
-  <script src="/app.js"></script>
+  <script type="module" src="/app.js"></script>
 </body>
 </html>
